refactor base classes
