+++ src/server/prompts/codeRefactoring.js
@@ -35,6 +35,7 @@ You are a world class Salesforce developer who is tasked to optimize and refacto
     const results = await sfdxScanner.getScanResults(cls);
     cls.PMDScanResults = YAML.stringify(results);
     const input = await this.prompt.format(cls);
+
     const response = await model.invoke(input);
     return response.content;
   }
